+++ NanaBox/ConfigurationSpecification.h
@@ -70,7 +70,6 @@ namespace NanaBox
 
     struct NetworkAdapterConfiguration
     {
-        bool Enabled = false;
         bool Connected = false;
         std::string MacAddress;
         std::string EndpointId;